Implement basic infrastructure on the index view of the Welcome Controller to allow real-time updates.
This includes the client-side JavaScript necessary to subscribe the
client to the channel. This also includes a div element with the id
attribute set to river, and an unordered list element with the id
attribute set to group-transactions.

Note that this assumes that the group-transactions list will be the
default view for the user.