Initial factory file.
Note that this factory file only defines factories for the Group, User,
and Request resources, and excludes the Vote and Transaction resources.
This was done due to the complexity of these associations. These
factories will be included in alter releases.

Also note that we can create requests in different states by calling
request_with_neighbor and fulfilled_request.